source/requirements.txt
deep_translator==1.11.4
groq==0.11.0
langchain==0.3.7
langchain_core==0.3.15
langdetect==1.0.9
langgraph==0.2.45
pinecone==5.3.1
python-dotenv==1.0.1
streamlit==1.39.0
tavily_python==0.5.0
typing_extensions==4.12.2
